+++ src/parsing/meta_interpret.c
@@ -6,14 +6,14 @@
 /*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/31 10:32:25 by aaugu             #+#    #+#             */
-/*   Updated: 2023/06/02 10:46:16 by aaugu            ###   ########.fr       */
+/*   Updated: 2023/06/02 15:41:36 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/parsing_input.h"
-#include "../includes/parsing_meta.h"
-#include "../includes/print_error.h"
-#include "../includes/minishell.h"
+#include "../../includes/parsing_input.h"
+#include "../../includes/parsing_meta.h"
+#include "../../includes/print_error.h"
+#include "../../includes/minishell.h"
 
 void	meta_interpret(t_token *t, char **env, int env_size)
 {
