+++ src/parsing/parsing-heredoc/state_dollar.c
@@ -1,0 +1,46 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   state_dollar.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/06/02 15:30:40 by aaugu             #+#    #+#             */
+/*   Updated: 2023/06/02 15:46:22 by aaugu            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+// #include <errno.h>
+// #include "../../../includes/parsing_meta_state_machine.h"
+// #include "../../../includes/minishell.h"
+// #include "../../../includes/print_error.h"
+// #include "../../../libft/libft.h"
+
+// void	state_dollar(t_m_fsm *fsm, t_meta **meta_strs, char c)
+// {
+// 	if (c == '?' || c == '$')
+// 		fsm->current_state = idle;
+// 	if (c == '?')
+// 	{
+// 		free(fsm->buf);
+// 		fsm->buf = ft_itoa(g_exit_code);
+// 		if (!fsm->buf)
+// 		{
+// 			parsing_error_meta(&(fsm->current_state));
+// 			return ;
+// 		}
+// 		finish_buf_meta(fsm, meta_strs);
+// 	}
+// 	else if (c == '$')
+// 		init_meta_state_machine(fsm);
+// 	else if (c == '\0')
+// 	{
+// 		finish_buf_meta(fsm, meta_strs);
+// 		fsm->current_state = stop;
+// 	}
+// 	else
+// 	{
+// 		add_to_buf_meta(fsm, c);
+// 		fsm->current_state = chars;
+// 	}
+// }
