+++ src/parsing/parsing-input/actions_finish_utils.c
@@ -6,11 +6,11 @@
 /*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/23 15:30:15 by aaugu             #+#    #+#             */
-/*   Updated: 2023/05/23 15:30:56 by aaugu            ###   ########.fr       */
+/*   Updated: 2023/06/02 15:39:11 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/parsing_input_state_machine.h"
+#include "../../../includes/parsing_input_state_machine.h"
 
 /* Combination of finish_buf() and add_to_buff() */
 void	finish_add(t_fsm *fsm, t_token **tokens, char c)
