+++ src/parsing/parsing-input/parsing_input.c
@@ -6,14 +6,14 @@
 /*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/02 11:18:11 by aaugu             #+#    #+#             */
-/*   Updated: 2023/05/23 16:00:39 by aaugu            ###   ########.fr       */
+/*   Updated: 2023/06/02 15:39:33 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
-#include "../../includes/parsing_input.h"
-#include "../../includes/parsing_input_state_machine.h"
-#include "../../libft/libft.h"
+#include "../../../includes/parsing_input.h"
+#include "../../../includes/parsing_input_state_machine.h"
+#include "../../../libft/libft.h"
 
 /* Parsing to retrieve each separate elements (tokens) of user input as an 
 element of the command and store them in a chained list. */
