+++ src/parsing/parsing-input/state_idle.c
@@ -1,39 +1,44 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   state_pipe.c                                       :+:      :+:    :+:   */
+/*   state_idle.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/05 10:42:25 by aaugu             #+#    #+#             */
-/*   Updated: 2023/05/31 16:32:26 by aaugu            ###   ########.fr       */
+/*   Created: 2023/05/05 13:25:13 by aaugu             #+#    #+#             */
+/*   Updated: 2023/06/02 15:39:41 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/parsing_input_state_machine.h"
+#include <stdbool.h>
+#include <stdio.h>
+#include "../../../includes/parsing_input_state_machine.h"
 
 /* Will set current state according to char and decide which action to perform
 if needed */
-void	state_pipe(t_fsm *fsm, t_token **tokens, char c)
+void	state_idle(t_fsm *fsm, t_token **tokens, char c)
 {
-	if (c == ' ' || c == '\'' || c == '\"')
-		finish_buf(fsm, tokens, c);
-	if (c == '<' || c == '>')
+	if (c == '<' || c == '>' || c == '|')
 		finish_add(fsm, tokens, c);
+	if (c == '\'' || c == '\"')
+		fsm->quotes = true;
 	if (c == '<')
-		fsm->current_state = less_than;
+		change_state_and_type(fsm, less_than, redir_in);
 	else if (c == '>')
-		fsm->current_state = greater_than;
+		change_state_and_type(fsm, greater_than, redir_out);
 	else if (c == '|')
-		parsing_error(fsm, &c);
+		change_state_and_type(fsm, s_pipe, t_pipe);
 	else if (c == '\'')
+	{
+		fsm->meta = false;
 		change_state_quotes_true(fsm, quote_s);
+	}
 	else if (c == '\"')
 		change_state_quotes_true(fsm, quote_d);
-	else if (c == ' ')
-		fsm->current_state = idle;
 	else if (c == '\0')
-		parsing_error(fsm, "newline");
+		finish_stop(fsm, tokens, c);
+	else if (c == ' ')
+		finish_buf(fsm, tokens, c);
 	else
-		finish_add_idle(fsm, tokens, c);
+		add_to_buf(fsm, c);
 }
