+++ src/parsing/parsing-input/state_machine.c
@@ -6,15 +6,15 @@
 /*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/05 11:54:32 by aaugu             #+#    #+#             */
-/*   Updated: 2023/05/26 13:52:15 by aaugu            ###   ########.fr       */
+/*   Updated: 2023/06/02 15:39:52 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include <stdbool.h>
-#include "../../includes/parsing_input_state_machine.h"
-#include "../../includes/parsing_input.h"
-#include "../../libft/libft.h"
+#include "../../../includes/parsing_input_state_machine.h"
+#include "../../../includes/parsing_input.h"
+#include "../../../libft/libft.h"
 
 t_fsm	create_state_machine(char *input);
 void	execute_state_machine(t_fsm *fsm, t_token **tokens, char c);
