+++ src/parsing/parsing-input/state_machine_utils.c
@@ -1,25 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parsing_error_meta.c                               :+:      :+:    :+:   */
+/*   state_machine_utils.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/26 11:11:22 by aaugu             #+#    #+#             */
-/*   Updated: 2023/05/26 15:21:35 by aaugu            ###   ########.fr       */
+/*   Created: 2023/05/16 11:22:46 by aaugu             #+#    #+#             */
+/*   Updated: 2023/06/02 15:39:47 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
-#include <errno.h>
-#include "../../includes/parsing_meta_state_machine.h"
-#include "../../includes/minishell.h"
+#include <stdbool.h>
+#include "../../../includes/parsing_input_state_machine.h"
+#include "../../../libft/libft.h"
 
-/* Handles error in state machine : Delivers error message and set state of 
-state machine to error */
-void	parsing_error_meta(t_state *current_state)
+/* Initialize start values of t_fsm state machine */
+void	init_state_machine(t_fsm *fsm)
 {
-	printf("minishell: malloc() failed: %s\n", strerror(errno));
-	g_exit_code = errno;
-	*current_state = error;
+	if (fsm->buf)
+		free(fsm->buf);
+	fsm->buf = (char *)ft_calloc(fsm->input_size, sizeof(char));
+	if (!fsm->buf)
+		parsing_error(fsm, 0);
+	fsm->buf_size = 0;
+	fsm->meta = true;
 }
