+++ src/parsing/parsing-input/state_pipe.c
@@ -1,44 +1,39 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   state_idle.c                                       :+:      :+:    :+:   */
+/*   state_pipe.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/05 13:25:13 by aaugu             #+#    #+#             */
-/*   Updated: 2023/05/31 16:38:11 by aaugu            ###   ########.fr       */
+/*   Created: 2023/05/05 10:42:25 by aaugu             #+#    #+#             */
+/*   Updated: 2023/06/02 15:39:54 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdbool.h>
-#include <stdio.h>
-#include "../../includes/parsing_input_state_machine.h"
+#include "../../../includes/parsing_input_state_machine.h"
 
 /* Will set current state according to char and decide which action to perform
 if needed */
-void	state_idle(t_fsm *fsm, t_token **tokens, char c)
+void	state_pipe(t_fsm *fsm, t_token **tokens, char c)
 {
-	if (c == '<' || c == '>' || c == '|')
+	if (c == ' ' || c == '\'' || c == '\"')
+		finish_buf(fsm, tokens, c);
+	if (c == '<' || c == '>')
 		finish_add(fsm, tokens, c);
-	if (c == '\'' || c == '\"')
-		fsm->quotes = true;
 	if (c == '<')
-		change_state_and_type(fsm, less_than, redir_in);
+		fsm->current_state = less_than;
 	else if (c == '>')
-		change_state_and_type(fsm, greater_than, redir_out);
+		fsm->current_state = greater_than;
 	else if (c == '|')
-		change_state_and_type(fsm, s_pipe, t_pipe);
+		parsing_error(fsm, &c);
 	else if (c == '\'')
-	{
-		fsm->meta = false;
 		change_state_quotes_true(fsm, quote_s);
-	}
 	else if (c == '\"')
 		change_state_quotes_true(fsm, quote_d);
-	else if (c == '\0')
-		finish_stop(fsm, tokens, c);
 	else if (c == ' ')
-		finish_buf(fsm, tokens, c);
+		fsm->current_state = idle;
+	else if (c == '\0')
+		parsing_error(fsm, "newline");
 	else
-		add_to_buf(fsm, c);
+		finish_add_idle(fsm, tokens, c);
 }
