+++ src/parsing/parsing-meta/parsing_error_meta.c
@@ -1,25 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   state_machine_meta_utils.c                         :+:      :+:    :+:   */
+/*   parsing_error_meta.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: aaugu <aaugu@student.42lausanne.ch>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/23 16:31:28 by aaugu             #+#    #+#             */
-/*   Updated: 2023/05/26 13:23:17 by aaugu            ###   ########.fr       */
+/*   Created: 2023/05/26 11:11:22 by aaugu             #+#    #+#             */
+/*   Updated: 2023/06/02 15:38:44 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
-#include "../../includes/parsing_meta_state_machine.h"
-#include "../../libft/libft.h"
+#include <stdio.h>
+#include <errno.h>
+#include "../../../includes/parsing_meta_state_machine.h"
+#include "../../../includes/minishell.h"
 
-void	init_meta_state_machine(t_m_fsm *fsm)
+/* Handles error in state machine : Delivers error message and set state of 
+state machine to error */
+void	parsing_error_meta(t_state *current_state)
 {
-	if (fsm->buf)
-		free(fsm->buf);
-	fsm->buf = (char *)ft_calloc(fsm->len, sizeof(char));
-	if (!fsm->buf)
-		parsing_error_meta(&(fsm->current_state));
-	fsm->buf_size = 0;
+	printf("minishell: malloc() failed: %s\n", strerror(errno));
+	g_exit_code = errno;
+	*current_state = error;
 }
